move hdd string processing functions to `util.c`
