+++ include/util.h
@@ -23,3 +23,19 @@ int loadIRXFile(char *path, u32 arg_len, const char *args, int *mod_res);
 int get_CNF_string(char **CNF_p_p,
                    char **name_p_p,
                    char **value_p_p);
+
+
+/**
+ * @brief  method returns true if it can extract needed info from path, otherwise false.
+ * In case of true, it also updates mountString, mountPoint and newCWD parameters
+ * It splits path by ":", and requires a minimum of 3 elements
+ * @example if path = hdd0:__common:pfs:/retroarch/ then: mountString = "pfs:", mountPoint = "hdd0:__common", newCWD = pfs:/retroarch/
+ * @param path input parameter with full hdd path (`hdd0:__common:pfs:/retroarch/`)
+ * @param mountString pointer to char* wich will contain pfs mountpoint (`pfs:`)
+ * @param mountPoint returns the path of mounted partition (`hdd0:__common`)
+ * @param newCWD returns the path to the file as pfs mount point string
+ * @return true on success
+*/
+int getMountInfo(char *path, char *mountString, char *mountPoint, char *newCWD);
+
+char** str_split(char* a_str, const char a_delim);
