+++ src/main.c
@@ -27,7 +27,6 @@
 #include <fileXio_rpc.h>
 #include <hdd-ioctl.h>
 #include <io_common.h>
-#include <assert.h>
 #endif
 
 #include "debugprintf.h"
@@ -693,7 +692,7 @@ int MountParty(const char* path)
     char* BUF = NULL;
     BUF = strdup(path);
     char MountPoint[40];
-    if (getMountInfo(BUF, NULL, MountPoint, NULL, NULL))
+    if (getMountInfo(BUF, NULL, MountPoint, NULL))
     {
         mnt(MountPoint);
         if (BUF != NULL)
@@ -724,105 +723,6 @@ int mnt(const char* path)
         }
     } else DPRINTF("mount successfull on first attemp\n");
     return 0;
-}
-
-//By fjtrujy
-static char** str_split(char* a_str, const char a_delim)
-{
-    char** result    = 0;
-    size_t count     = 0;
-    char* tmp        = a_str;
-    char* last_comma = 0;
-    char delim[2];
-    delim[0] = a_delim;
-    delim[1] = 0;
-
-    /* Count how many elements will be extracted. */
-    while (*tmp)
-    {
-        if (a_delim == *tmp)
-        {
-            count++;
-            last_comma = tmp;
-        }
-        tmp++;
-    }
-
-    /* Add space for trailing token. */
-    count += last_comma < (a_str + strlen(a_str) - 1);
-
-    /* Add space for terminating null string so caller
-       knows where the list of returned strings ends. */
-    count++;
-
-    result = malloc(sizeof(char*) * count);
-
-    if (result)
-    {
-        size_t idx  = 0;
-        char* token = strtok(a_str, delim);
-
-        while (token)
-        {
-            assert(idx < count);
-            *(result + idx++) = strdup(token);
-            token = strtok(0, delim);
-        }
-        assert(idx == count - 1);
-        *(result + idx) = 0;
-    }
-
-    return result;
-}
-
-/**
- * @brief  method returns true if it can extract needed info from path, otherwise false.
- * In case of true, it also updates mountString, mountPoint and newCWD parameters
- * It splits path by ":", and requires a minimum of 3 elements
- * Example: if path = hdd0:__common:pfs:/retroarch/ then
- * mountString = "pfs:"
- * mountPoint = "hdd0:__common"
- * newCWD = pfs:/retroarch/
- * return true
-*/
-int getMountInfo(char *path, char *mountString, char *mountPoint, char *newCWD)
-{
-    int expected_items = 4;
-    int i = 0;
-    char *items[expected_items];
-    char** tokens = str_split(path, ':');
-
-    if (!tokens)
-        return 0;
-    
-    for (i = 0; *(tokens + i); i++) {
-        if (i < expected_items) {
-            items[i] = *(tokens + i);
-        } else {
-            free(*(tokens + i));    
-        }
-    }
-
-    if (i < 3 )
-        return 0;
-    
-    if (mountPoint != NULL)
-        sprintf(mountPoint, "%s:%s", items[0], items[1]);
-
-    if (mountString != NULL)
-        sprintf(mountString, "%s:", items[2]);
-
-    if (newCWD != NULL)
-        sprintf(newCWD, "%s:%s", items[2], i > 3 ? items[3] : "");
-    
-    free(items[0]);
-    free(items[1]);
-    free(items[2]);
-
-    if (i > 3)
-        free(items[3]);
-
-    return 1;
 }
 #endif
